Inclusão de acompanhamento de estoque

+++ app/controllers/item_orcamentos_controller.rb
@@ -15,21 +15,24 @@ class ItemOrcamentosController < ApplicationController
         @item_orcamento = ItemOrcamento.new
         @produtos = Produto.all
         @pedido_orcamento = PedidoOrcamento.find(params[:pedido_orcamento_id])
-        puts "Entrei no NEW"
     end
 
     def create
-        puts "Entrei no CREATE1"
         @item_orcamento = ItemOrcamento.new(item_orcamento_params)
         @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
         
-        if @item_orcamento.save 
-            @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
-            redirect_to pedido_orcamento_item_orcamento_path(@pedido_orcamento, @item_orcamento)
-            puts @item_orcamento
+        if remover_produto(@item_orcamento.produto_id, @item_orcamento.quantidade)
+            if @item_orcamento.save 
+                @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
+                redirect_to pedido_orcamento_item_orcamento_path(@pedido_orcamento, @item_orcamento)
+                puts @item_orcamento
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
+            puts "OPS"
         end
+        
     end
 
     def edit
@@ -39,28 +42,58 @@ class ItemOrcamentosController < ApplicationController
     end
 
     def update
-        @item_orcamento = ItemOrcamento.find(params[:id])        
-        if @item_orcamento.update(item_orcamento_params)
-            @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
-            redirect_to pedido_orcamento_item_orcamento_path(@pedido_orcamento, @item_orcamento)
-            puts @item_orcamento
+        @item_orcamento = ItemOrcamento.find(params[:id])  
+        @quantidade_anterior = @item_orcamento.quantidade
+        @quantidade_atual = params[:item_orcamento][:quantidade]
+        
+        if adicionar_produto(@item_orcamento.produto_id, @quantidade_anterior) && remover_produto(@item_orcamento.produto_id, @quantidade_atual.to_i)
+            if @item_orcamento.update(item_orcamento_params)
+                @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
+                redirect_to pedido_orcamento_item_orcamento_path(@pedido_orcamento, @item_orcamento)
+                puts @item_orcamento
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
-        end
+            puts "OPS"
+        end 
     end
 
     def destroy
         @item_orcamento = ItemOrcamento.find(params[:id])
         @pedido_orcamento = PedidoOrcamento.find(@item_orcamento.pedido_orcamento_id)
-        if @item_orcamento.destroy
-            redirect_to pedido_orcamento_item_orcamentos_path(@pedido_orcamento)
+        
+        if adicionar_produto(@item_orcamento.produto_id, @item_orcamento.quantidade)
+            if @item_orcamento.destroy
+                redirect_to pedido_orcamento_item_orcamentos_path(@pedido_orcamento)
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
-        end
+            puts "OPS"
+        end       
     end
 
     private
     def item_orcamento_params
         params.require(:item_orcamento).permit(:produto_id, :pedido_orcamento_id, :quantidade)
+    end
+
+    private
+    def adicionar_produto(produto_id, quantidade)
+        @produto = Produto.find(produto_id)
+        if @produto.update(quantidade: @produto.quantidade + quantidade)
+            return true
+        end
+        false
+    end
+
+    private
+    def remover_produto(produto_id, quantidade)
+        @produto = Produto.find(produto_id)
+        if @produto.update(quantidade: @produto.quantidade - quantidade)
+            return true
+        end
+        false
     end
 end
