+++ app/controllers/item_pedidos_controller.rb
@@ -21,14 +21,19 @@ class ItemPedidosController < ApplicationController
     def create
         @item_pedido = ItemPedido.new(item_pedido_params)
         @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
-        
-        if @item_pedido.save 
-            @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
-            redirect_to pedido_supplier_item_pedido_path(@pedido_supplier, @item_pedido)
-            puts @item_pedido
+
+        if adicionar_produto(@item_pedido.produto_id, @item_pedido.quantidade)
+            if @item_pedido.save 
+                @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
+                redirect_to pedido_supplier_item_pedido_path(@pedido_supplier, @item_pedido)
+                puts @item_pedido
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
+            puts "OPS"
         end
+        
     end
 
     def edit
@@ -38,28 +43,58 @@ class ItemPedidosController < ApplicationController
     end
 
     def update
-        @item_pedido = ItemPedido.find(params[:id])        
-        if @item_pedido.update(item_pedido_params)
-            @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
-            redirect_to pedido_supplier_item_pedido_path(@pedido_supplier, @item_pedido)
-            puts @item_pedido
+        @item_pedido = ItemPedido.find(params[:id]) 
+        @quantidade_anterior = @item_pedido.quantidade
+        @quantidade_atual = params[:item_pedido][:quantidade]
+        
+        if remover_produto(@item_pedido.produto_id, @quantidade_anterior) && adicionar_produto(@item_pedido.produto_id, @quantidade_atual.to_i)
+            if @item_pedido.update(item_pedido_params)
+                @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
+                redirect_to pedido_supplier_item_pedido_path(@pedido_supplier, @item_pedido)
+                puts @item_pedido
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
-        end
+            puts "OPS"
+        end        
     end
 
     def destroy
         @item_pedido = ItemPedido.find(params[:id])
         @pedido_supplier = PedidoSupplier.find(@item_pedido.pedido_supplier_id)
-        if @item_pedido.destroy
-            redirect_to pedido_supplier_item_pedidos_path(@pedido_supplier)
+
+        if remover_produto(@item_pedido.produto_id, @item_pedido.quantidade)
+            if @item_pedido.destroy
+                redirect_to pedido_supplier_item_pedidos_path(@pedido_supplier)
+            else
+                puts "Deu ruim"
+            end
         else
-            puts "Deu ruim"
-        end
+            puts "OPS"
+        end    
     end
 
     private
     def item_pedido_params
         params.require(:item_pedido).permit(:produto_id, :pedido_supplier_id, :quantidade)
+    end
+
+    private
+    def adicionar_produto(produto_id, quantidade)
+        @produto = Produto.find(produto_id)
+        if @produto.update(quantidade: @produto.quantidade + quantidade)
+            return true
+        end
+        false
+    end
+
+    private
+    def remover_produto(produto_id, quantidade)
+        @produto = Produto.find(produto_id)
+        if @produto.update(quantidade: @produto.quantidade - quantidade)
+            return true
+        end
+        false
     end
 end
